[fix]: date 유틸함수를 한 파일에서 관리
